src/main/java/com/kaven/weixinsell/service/impl/ProductCategoryServiceImpl.java
package com.kaven.weixinsell.service.impl;

import com.kaven.weixinsell.dataobject.ProductCategory;
import com.kaven.weixinsell.repository.ProductCategoryRepository;
import com.kaven.weixinsell.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory product) {
        return repository.save(product);
    }
}
